backend/requirements.txt
flask
flask-cors
pyrebase4
google-generativeai
python-dotenv
